Yêu cầu 8 9 10 11 13 15

+++ QuanLySinhVien/src/main/java/myproject/QuanLySinhVien/entities/DiemSoDAO.java
@@ -53,9 +53,14 @@ public class DiemSoDAO {
 		Session session=HibernateUtil.getSessionFactory().openSession();
 		Transaction transaction=session.beginTransaction();
 		try {
-				session.delete(ds);
+				if(session.get(DiemSo.class, ds.getId()) != null)
+				{
+					session.clear();
+					session.delete(ds);
+				}else
+					throw new Exception("Student not join this subject");
 				transaction.commit();
-		}catch(HibernateException e){
+		}catch(Exception e){
 			transaction.rollback();
 			e.printStackTrace();
 			return false;
@@ -66,5 +71,37 @@ public class DiemSoDAO {
 		
 		return true;
 		
+	}
+
+
+	public static List<DiemSo> getBySubject(String Mon){
+		Session session=HibernateUtil.getSessionFactory().openSession();
+		List<DiemSo> ds = null;
+		try {
+			String hql = "select ds from DiemSo ds where MaMh=:Mon";
+			ds = session.createQuery(hql).setParameter("Mon", Mon).getResultList();
+		} catch (HibernateException ex) {
+			System.err.println(ex);
+		} finally {
+			session.close();
+		}
+
+		return ds;
+	}
+
+
+	public static List<DiemSo> getBySinhVien(int mssv) {
+		Session session=HibernateUtil.getSessionFactory().openSession();
+		List<DiemSo> ds = null;
+		try {
+			String hql = "select ds from DiemSo ds where Mssv=:mssv";
+			ds = session.createQuery(hql).setParameter("Mssv", mssv).getResultList();
+		} catch (HibernateException ex) {
+			System.err.println(ex);
+		} finally {
+			session.close();
+		}
+
+		return ds;
 	}
 }
